Fixes for busybox head, tail and awk
Busybox based OS stacks like Alpine have a simpler parser so a few of
the commands in the original script do no work. Specifically for head
and tail, the options must be ahead of the file. In awk, the  prefix
increment breaks the string concat so this converts it to a postfix
increment and rebases the sequence from 0.